add hf support
