+++ nemo_curator/classifiers/domain.py
@@ -15,14 +15,12 @@ import os
 from dataclasses import dataclass
 
 os.environ["RAPIDS_NO_INITIALIZE"] = "1"
-import torch
-import torch.nn as nn
 from crossfit.backend.torch.hf.model import HFModel
-from huggingface_hub import PyTorchModelHubMixin
-from transformers import AutoConfig, AutoModel, AutoTokenizer
+from transformers import AutoConfig, AutoTokenizer
 
 from nemo_curator.classifiers.base import (
     DistributedDataClassifier,
+    HFCustomModel,
     _run_classifier_helper,
 )
 from nemo_curator.datasets import DocumentDataset
@@ -35,32 +33,6 @@ class DomainModelConfig:
     model = "microsoft/deberta-v3-base"
     fc_dropout = 0.2
     max_len = 512
-
-
-class HFCustomModel(nn.Module, PyTorchModelHubMixin):
-    def __init__(self, config: dataclass):
-        super(HFCustomModel, self).__init__()
-        self.model = AutoModel.from_pretrained(config["base_model"])
-        self.dropout = nn.Dropout(config["fc_dropout"])
-        self.fc = nn.Linear(self.model.config.hidden_size, len(config["id2label"]))
-
-    def _forward(self, batch):
-        features = self.model(
-            batch["input_ids"], batch["attention_mask"]
-        ).last_hidden_state
-        dropped = self.dropout(features)
-        outputs = self.fc(dropped)
-        return torch.softmax(outputs[:, 0, :], dim=1)
-
-    def forward(self, batch):
-        if self.autocast:
-            with torch.autocast(device_type="cuda"):
-                return self._forward(batch)
-        else:
-            return self._forward(batch)
-
-    def set_autocast(self, autocast):
-        self.autocast = autocast
 
 
 class DomainModel(HFModel):
